requirements.txt
python-telegram-bot==21.0.1
google-generativeai==0.8.3
gspread==6.1.2
oauth2client==4.1.3
python-dotenv==1.0.1
APScheduler==3.10.4
aiohttp==3.9.1
